+++ libs/cardano-ledger-canonical-state/src/Cardano/Ledger/CanonicalState/Namespace/EntitiesStakePools/V0/CDDL.hs
@@ -57,6 +57,7 @@ stake_pool_state =
       , "relays" ==> arr [0 <+ a relay]
       , "deposit" ==> coin
       , "metadata" ==> pool_metadata / VNil
+      , "bls_key" ==> bls_key_rule / VNil
       , "account_id" ==> account_id
       , "delegators" ==> set credential
       ]
@@ -67,6 +68,7 @@ stake_pool_params =
     =:= mp
       [ "id" ==> pool_keyhash
       , "vrf" ==> vrf_keyhash
+      , "bls_key" ==> bls_key_rule / VNil
       , "cost" ==> coin
       , "margin" ==> unit_interval
       , "owners" ==> set staking_keyhash
@@ -75,3 +77,6 @@ stake_pool_params =
       , "metadata" ==> pool_metadata / VNil
       , "account_address" ==> address
       ]
+
+bls_key_rule :: Rule
+bls_key_rule = "bls_key" =:= arr [a VBytes, a VBytes]